+++ src/SDKs/Azure.ApplicationModel.Configuration/data-plane/Azure.ApplicationModel.Configuration/ConfigurationSettingParser.cs
@@ -136,31 +136,5 @@ namespace Azure.ApplicationModel.Configuration
             return true;
         }
     }
-
-    // TODO (pri 3): CoreFx will soon have a type like this. We should remove this one then.
-    internal class FixedSizedBufferWriter : IBufferWriter<byte>
-    {
-        private readonly byte[] _buffer;
-        private int _count;
-
-        public FixedSizedBufferWriter(byte[] buffer)
-        {
-            _buffer = buffer;
-        }
-
-        public Memory<byte> GetMemory(int minimumLength = 0) => _buffer.AsMemory(_count);
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public Span<byte> GetSpan(int minimumLength = 0) => _buffer.AsSpan(_count);
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Advance(int bytes)
-        {
-            _count += bytes;
-            if (_count > _buffer.Length)
-            {
-                throw new InvalidOperationException("Cannot advance past the end of the buffer.");
-            }
-        }
-    }
 }
+
